+++ src/javascript/constructors/dialog.js
@@ -1,13 +1,16 @@
 import DialogBody from '../components/dialog/dialog'
 import WindowBackground  from '../components/window/background'
 import {puffin} from '@mkenzo_8/puffin'
-import {Titles,Text } from '@mkenzo_8/puffin-drac'
+import { Titles, Text, Button } from '@mkenzo_8/puffin-drac'
 
 function Dialog({
     title = 'Title',
-    content = ""
+    content = "",
+    component,
+    buttons = []
 }){
     const randomSelector = Math.random()
+    const computedMethods = {...buttons.map(btn=>btn.action)}
     const DialogComp = puffin.element(`
         <div id="${randomSelector}" class="${puffin.style.css`
             &{
@@ -17,25 +20,46 @@ function Dialog({
                 top:50%;
                 left:50%;
                 transform:translate(-50%,-50%);
-
             }
         `}">
             <WindowBackground window="${randomSelector}"/>
             <DialogBody>
-                <H1>${title}</H1>
+                <H2>${title}</H2>
                 <Text>${content}</Text>
+                <div>
+                    ${(function(){
+                        let content = "";
+                        buttons.map(function(btn,index){
+                            content += `<Button click="$${index}" onclick="document.getElementById('${randomSelector}').remove()">${btn.label}</Button>`
+                        })
+                        return content
+                    })()}
+                </div>
             </DialogBody>
         </div>
     `,{
         components:{
             DialogBody,
             WindowBackground,
-            H1:Titles.h1,
-            Text
-
-        }
+            H2:Titles.h2,
+            Text,
+            Button
+        },
+        events:{
+            mounted(target){
+                if(component != null){
+                    puffin.render(component,target.children[1].children[1])
+                }
+            }
+        },
+        methods:computedMethods
     })
+
     
     puffin.render(DialogComp,document.getElementById("windows"))
+    return {
+        close:()=> document.getElementById(randomSelector).remove()
+    }
 }
+
 export default Dialog