+++ src/javascript/defaults/dialogs/about.js
@@ -1,0 +1,12 @@
+import Dialog from '../../constructors/dialog'
+import GravitonPackage from '../../../../package.json'
+
+function AboutDialog(){
+    new Dialog({
+        title:'About',
+        content:`Graviton v${GravitonPackage.version}`
+    })
+}
+
+
+export default AboutDialog