+++ src/javascript/utils/directory.parser.js
@@ -1,0 +1,15 @@
+import requirePath from './require'
+
+const path = requirePath("path")
+
+function parseDirectory(directory){
+    let nameFolder = path.basename(directory)
+
+    if(process.platform == "win32"){
+            nameFolder = path.basename(directory.replace(/\\/g,'\\\\'))
+    }
+
+    return nameFolder
+}
+
+export default parseDirectory