+++ src/javascript/utils/themeprovider.js
@@ -2,6 +2,7 @@ import {puffin} from '@mkenzo_8/puffin'
 import StaticConfig from 'StaticConfig'
 import Arctic from '../themes/arctic'
 import Night from '../themes/night'
+import ThemeRegistry from 'ThemeRegistry'
 
 StaticConfig.changed(function(){
     applyTheme(StaticConfig)
@@ -55,17 +56,8 @@ const ThemeProvider = new puffin.state({
 })
 
 function applyTheme(state){
-    switch(state.data.theme){
-        case "Arctic":
-            ThemeProvider.data = Arctic
-        break;
-        case "Night":
-            ThemeProvider.data = Night
-        break;
-    }
+    ThemeProvider.data = ThemeRegistry.registry.data.colorsSchemes[state.data.theme]
     ThemeProvider.triggerChange()
 }
-
-applyTheme(StaticConfig)
 
 export default ThemeProvider